Rename to ephys link
